pip-requirements.txt
pymongo==2.1.1
XenAPI==1.2
celery-with-redis==2.5
